qa/qc'ed group modification user exists/in group

+++ giftpal/groups.py
@@ -48,17 +48,33 @@ def mod_group(group_id, group, query_group):
         group.min_dollar_amount = request.form['min_dollar_amount']
 
     user = User.query.filter_by(username=request.form['modify_selected_user']).first()
-    if 'modify_selected_user' in request.form:
-        action_to_group = request.form['group_modification']
-        if action_to_group == "add" and user is not None:
-            new_user_group = UserGroup(user_id=user.id, group_id=query_group.id, is_admin=False)
-            db.session.add(new_user_group)
-        elif action_to_group == "delete":
-            deleted_user_group = UserGroup.query.filter_by(user_id=user.id, group_id=group_id).first()
-            db.session.delete(deleted_user_group)
-        elif action_to_group == "make_admin":
-            user_group = UserGroup.query.filter_by(user_id=user.id, group_id=group_id).first()
-            user_group.is_admin = True
+    if user is not None: 
+        if 'modify_selected_user' in request.form:
+            action_to_group = request.form['group_modification']
+            user_in_group_already = UserGroup.query.filter_by(user_id=user.id, group_id=group_id).first()
+            if action_to_group == "add":
+                if user_in_group_already is None:
+                    new_user_group = UserGroup(user_id=user.id, group_id=query_group.id, is_admin=False)
+                    db.session.add(new_user_group)
+                    flash('User successfully added to group!')
+                else: 
+                    flash('User is already in group!')
+            else:
+                if user_in_group_already is None:
+                    flash('User is not in group!')
+                else: 
+                    if action_to_group == "delete":
+                        deleted_user_group = UserGroup.query.filter_by(user_id=user.id, group_id=group_id).first()
+                        db.session.delete(deleted_user_group)
+                        flash('User successfully deleted from group!')
+                    elif action_to_group == "make_admin":
+                        user_group = UserGroup.query.filter_by(user_id=user.id, group_id=group_id).first()
+                        user_group.is_admin = True
+                        flash('User successfully made admin of group!')
+
+
+    else: 
+        flash('User does not exist!')
 
     db.session.commit()
